# backend/openapi.yaml
openapi: 3.1.0
info:
  title: MatchmakerAPI
  version: 1.0.0
  description: >
    REST API for Matchmaker. Authentication is handled via Discord OAuth2.
    After completing the OAuth2 flow, clients receive a short-lived JWT access token
    and a long-lived HttpOnly refresh token cookie.

    Production (behind Cloudflare): when `ORIGIN_VERIFY_SECRET` is set, requests must
    include a matching `X-Origin-Verify` header (injected by the edge Worker). Missing
    or wrong values return 403. `GET /health` is exempt for platform probes. Local
    Compose leaves the secret unset, so this gate is inactive in development.
...
paths:
...
  /auth/refresh:
    post:
      tags: [auth]
      summary: Refresh access token
      operationId: refreshToken
      description: >
        Reads the `refresh_token` HttpOnly cookie, validates it, and issues a new
        access token plus a rotated refresh token (old token is deleted).
      responses:
        "200":
          description: New access token issued and refresh token rotated
          headers:
            Set-Cookie:
              schema:
                type: string
              description: Rotated `refresh_token` and `auth_session` cookies
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccessTokenResponse"
        "401":
          description: Missing, invalid, or expired refresh token cookie
        "500":
          description: Token generation or storage failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from `/auth/complete` or `/auth/refresh`
  schemas:
    AccessTokenResponse:
      type: object
      required: [access_token]
      properties:
        access_token:
          type: string
          description: "Short-lived JWT for use in the `Authorization: Bearer` header"
          examples: ["eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."]
    ErrorResponse:
      type: object
      required: [status, message]
      properties:
        status:
          type: string
          enum: [error]
          examples: [error]
        message:
          type: string
          examples: ["Could not generate required tokens"]
...